Python/requirements.txt
pandas==2.3.0
matplotlib==3.10.3
fpdf2==2.8.3
